requirements.txt
discord.py
asyncpg
python-dotenv
google-genai
